requirements.txt
pip
setuptools
azure-core
azure-storage-blob
